extract source info
